Other/AdvanceHealthMapUsingHyperSpectralImage/API2/requirements.txt
requests
seaborn
opencv-python
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow==2.15.0
numpy==1.24.3
pillow==10.1.0
pandas==2.1.4
python-multipart==0.0.6
pydantic==2.5.2
matplotlib==3.8.2
scikit-image==0.22.0
scipy==1.11.4
opencv-python==4.8.1.78
python-jose[cryptography]==3.3.0
python-decouple==3.8
aiofiles==23.2.0
jinja2==3.1.2
